feat: adding findPeopleByLocation()

+++ src/main/java/site/wuct/scholars/repository/LocationsRepository.java
@@ -1,8 +1,8 @@
 package site.wuct.scholars.repository;
 
 import site.wuct.scholars.model.Location;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LocationsRepository extends JpaRepository<Location, Long> {
-    
+public interface LocationsRepository extends JpaRepository<Location, Integer> {
 }
