+++ src/main/java/site/wuct/scholars/repository/PeopleRepository.java
@@ -3,5 +3,5 @@ package site.wuct.scholars.repository;
 import site.wuct.scholars.model.Person;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PeopleRepository extends JpaRepository<Person, Long> {
+public interface PeopleRepository extends JpaRepository<Person, Integer> {
 }
